requirements.txt
reflex>=0.4.0
streamlit>=1.29.0
openai>=1.3.0
pygame>=2.5.2
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
asyncio-mqtt>=0.16.1
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=2.0.0
watchdog>=3.0.0
typing-extensions>=4.5.0
httpx>=0.25.0
jinja2>=3.1.2
colorama>=0.4.6
